Add components to shared module
