+++ lib/dfe/analytics/services/generic_checksum_calculator.rb
@@ -1,4 +1,5 @@
 require_relative '../shared/service_pattern'
+require_relative '../shared/checksum_query_components'
 
 module DfE
   module Analytics
@@ -7,6 +8,7 @@ module DfE
       # and order column in a generic database
       class GenericChecksumCalculator
         include ServicePattern
+        include ChecksumQueryComponents
 
         WHERE_CLAUSE_ORDER_COLUMNS = %w[CREATED_AT UPDATED_AT].freeze
 
@@ -44,28 +46,6 @@ module DfE
 
           table_ids = connection.execute(checksum_sql_query).pluck('id')
           [table_ids.count, Digest::MD5.hexdigest(table_ids.join)]
-        end
-
-        def build_select_and_order_clause(order_column, table_name_sanitized)
-          case order_column
-          when 'updated_at'
-            select_clause = "#{table_name_sanitized}.updated_at as updated_at_alias"
-            order_by_clause = "updated_at_alias ASC, #{table_name_sanitized}.ID ASC"
-          when 'created_at'
-            select_clause = "#{table_name_sanitized}.created_at as created_at_alias"
-            order_by_clause = "created_at_alias ASC, #{table_name_sanitized}.ID ASC"
-          else
-            select_clause = ''
-            order_by_clause = "#{table_name_sanitized}.ID ASC"
-          end
-
-          [select_clause, order_by_clause]
-        end
-
-        def build_where_clause(order_column, table_name_sanitized, checksum_calculated_at_sanitized)
-          return '' unless WHERE_CLAUSE_ORDER_COLUMNS.map(&:downcase).include?(order_column.downcase)
-
-          "WHERE #{table_name_sanitized}.#{order_column.downcase} < #{checksum_calculated_at_sanitized}"
         end
       end
     end
