+++ lib/dfe/analytics/services/postgres_checksum_calculator.rb
@@ -1,4 +1,5 @@
 require_relative '../shared/service_pattern'
+require_relative '../shared/checksum_query_components'
 
 module DfE
   module Analytics
@@ -7,6 +8,7 @@ module DfE
       # and order column in a PostgreSQL database
       class PostgresChecksumCalculator
         include ServicePattern
+        include ChecksumQueryComponents
 
         WHERE_CLAUSE_ORDER_COLUMNS = %w[CREATED_AT UPDATED_AT].freeze
 
@@ -44,28 +46,6 @@ module DfE
 
           result = connection.execute(checksum_sql_query).first
           [result['row_count'].to_i, result['checksum']]
-        end
-
-        def build_select_and_order_clause(order_column, table_name_sanitized)
-          case order_column.downcase
-          when 'updated_at'
-            select_clause = "#{table_name_sanitized}.updated_at as updated_at_alias, "
-            order_by_clause = 'ORDER BY CHECKSUM_TABLE.updated_at_alias ASC, CHECKSUM_TABLE.ID ASC'
-          when 'created_at'
-            select_clause = "#{table_name_sanitized}.created_at as created_at_alias, "
-            order_by_clause = 'ORDER BY CHECKSUM_TABLE.created_at_alias ASC, CHECKSUM_TABLE.ID ASC'
-          else
-            select_clause = ''
-            order_by_clause = 'ORDER BY CHECKSUM_TABLE.ID ASC'
-          end
-
-          [select_clause, order_by_clause]
-        end
-
-        def build_where_clause(order_column, table_name_sanitized, checksum_calculated_at_sanitized)
-          return '' unless WHERE_CLAUSE_ORDER_COLUMNS.map(&:downcase).include?(order_column.downcase)
-
-          "WHERE #{table_name_sanitized}.#{order_column.downcase} < #{checksum_calculated_at_sanitized}"
         end
       end
     end
